third_party/blink/web_tests/http/tests/websocket/permessage-deflate-invalid-parameter-expected.txt

Test whether WebSocket rejects invalid permessage-deflate parameters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing parameter: "x-foo"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "client_max_window_bits=7"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "client_max_window_bits=16"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "client_no_context_takeover=foo"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "client_max_window_bits=8; client_no_context_takeover; x-foo"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
